requirements.txt
psutil
pyyaml
fastapi
uvicorn
